fix: typeof point

+++ backend/server/models/grade.go
@@ -3,7 +3,7 @@ package models
 type SubjectGrade struct {
 	SubjectName string  `json:"subjectName,omitempty"` // 科目名
 	UnitNum     int     `json:"unitNum,omitempty"`     // 単位
-	Point       int     `json:"point,omitempty"`       // 点数
+	Point       float64 `json:"point,omitempty"`       // 点数
 	Gp          float64 `json:"gp,omitempty"`          // GP
 	ReportedAt  string  `json:"reportedAt,omitempty"`  // 報告日
 }
